refactor: move project card to ui components
