+++ project/components/sections/projects/projects-list.tsx
@@ -1,7 +1,7 @@
 "use server";
 
 import { CreateProjectButton } from "@/components/create-project-button";
-import ProjectCard from "@/components/project-card";
+import ProjectCard from "@/components/ui/project-card";
 import { getProjects } from "@/lib/actions/projects";
 
 export default async function ProjectsList() {
